libs/aegis_auth/pyproject.toml
[project]
name = "aegis-auth"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pyjwt>=2.8", "sqlalchemy>=2.0", "httpx>=0.27"]
